+++ lib/src/entry.dart
@@ -1,10 +1,9 @@
 import 'package:dio/dio.dart';
 
-import '../api_manager.dart';
 import 'api_manager.dart';
 
-/// Created by Taohid on 02, March, 2020
+/// Created by Taohid on 01, March, 2020
 /// Email: taohid32@gmail.com
 
 ApiManager createApiManager(BaseOptions _baseOptions) =>
-    ApiManagerSrc(baseOptions: _baseOptions);
+    throw UnsupportedError('');