+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.4 - 0.1.6]
+Fixed list serialization issue
+added multipart file upload
+
+## [0.1.3]
+Response body wrapper added
+Some Bug Fixed
+
+## [0.1.0 - 0.1.2]
+Some Bug Fixed
+
 ## [0.0.2 - 0.0.3]
 Some Bug Fixed